tb_axi_lite_regs: Remove hardcoded assertion.
At simulation start the tb does one write followed by a read to a known writeable location. With different randomization seed values there is the possibility that the location gets overwritten in the meantime. The hardcoded assertion then fails. The assertion can be removes safely as the read data is checked against the model in proc_check_read_data at the same time. There the model is consistant with the expected behaviour.