axi_id_remap: Prevent "illegal" indexing with `MaxUniqInpIds == 1`
If `MaxUniqInpIds == 1`, the index of `table_q` is zero bit wide.
Before this commit, there were two cases where a one-bit signal was used
to index the table.  Although the result was never used, this was a
transiently an illegal index and caused annoying warnings.  This commit
handles the special case of `MaxUniqInpIds` with a safe static index
into `table_q`.